+++ CHANGELOG.md
@@ -5,14 +5,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [3.21.2005.0]
+## [3.21.2005.0] (not yet released)
 
 ### Added
 
 ### Changed
-- Invoke-PnPSearchQuery: Allow SelectProperties to take a comma separated string as well as an array
+- Fixed using `Connect-PnPOnline -AppId <appid> -AppSecret <appsecret> -AADDomain` not actually authenticating to Microsoft Graph [PR #2624](https://github.com/pnp/PnP-PowerShell/pull/2624)
 
 ### Contributors
+- Koen Zomers [koenzomers]
 
 ## [3.20.2004.0]